pyproject.toml
[tool.poetry]
name = "role-based_app"
version = "1.3.2"
description = "Backend for role-based app"
authors = []

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.9.2"
SQLAlchemy = "^1.4.41"
SQLAlchemy-Utils = "^0.41.0"
marshmallow = "^3.19.0"
passlib = "^1.7.3"
mangum = "^0.16.0"
fastapi = "^0.68.1"
email-validator = "^2.0.0"
pyjwt = "^2.5.0"
phonenumbers = "^8.13.11"
alembic = "^1.8.1"
python-decouple = "^3.6"
uvicorn="^0.21.1"
boto3 = "^1.26.143"
botocore = "^1.29.147"
jinja2 = "^3.1.1"
moto = "^4.1.11"
flake8 = "^6.0.0"

[tool.poetry.dev-dependencies]
pylint = "^2.14.1"
pytest = "^7.1.2"
Faker = "^13.12.1"
pytest-mock = "^3.8.1"
black = "^22.6.0"
python-dotenv = "^0.21.0"
psycopg2-binary = "^2.9.5"
httpx = "^0.19.0"
requests = "^2.30.0"
pytest-asyncio = "^0.20.2"
starlette = "^0.14.2"
anyio = "^3.3.0"

[tool.black]
line-length = 99

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
